+++ CHANGELOG.md
@@ -6,6 +6,4 @@
 - VMA-01: Create basic structure, README, CHANGELOG and Makefile
 - VMA-02: Implement "List" endpoint (GET: /users) with configurations module and database
 - VMA-03: Implement "Get" endpoint (GET: /users/{id})
-
-
-
+- VMA-04: Implement "Create" endpoint (POST: /users)